backend/requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.13
psycopg2-binary
django-cors-headers
gunicorn>=20.1.0
dj-rest-auth
django-filter
djangorestframework-simplejwt
openpyxl>=3.0.9
pandas>=1.3.4
PyYAML
uritemplate
drf-yasg
redis>=3.5.3,<4.0.0
celery
django-simple-history
django-solo
django-environ
